Move to "flick_electric" as domain
