+++ homeassistant/components/flick_electric/manifest.json
@@ -1,8 +1,8 @@
 {
-  "domain": "flickelectric",
+  "domain": "flick_electric",
   "name": "Flick Electric",
   "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/flickelectric/",
+  "documentation": "https://www.home-assistant.io/integrations/flick_electric/",
   "requirements": [
     "PyFlick==0.0.2"
   ],
