+++ homeassistant/components/flick_electric/sensor.py
