+++ homeassistant/components/flick_electric/strings.json
