+++ homeassistant/components/flick_electric/translations/en.json
