+++ tests/components/flick_electric/test_config_flow.py
@@ -5,7 +5,7 @@ from asynctest import patch
 from pyflick.authentication import AuthException
 
 from homeassistant import config_entries, data_entry_flow, setup
-from homeassistant.components.flickelectric.const import DOMAIN
+from homeassistant.components.flick_electric.const import DOMAIN
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 
 from tests.common import MockConfigEntry
@@ -29,12 +29,12 @@ async def test_form(hass):
     assert result["errors"] == {}
 
     with patch(
-        "homeassistant.components.flickelectric.config_flow.SimpleFlickAuth.async_get_access_token",
+        "homeassistant.components.flick_electric.config_flow.SimpleFlickAuth.async_get_access_token",
         return_value="123456789abcdef",
     ), patch(
-        "homeassistant.components.flickelectric.async_setup", return_value=True
+        "homeassistant.components.flick_electric.async_setup", return_value=True
     ) as mock_setup, patch(
-        "homeassistant.components.flickelectric.async_setup_entry", return_value=True,
+        "homeassistant.components.flick_electric.async_setup_entry", return_value=True,
     ) as mock_setup_entry:
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"], CONF,
@@ -54,12 +54,12 @@ async def test_form_duplicate_login(hass):
         domain=DOMAIN,
         data=CONF,
         title="Flick Electric: test-username",
-        unique_id="flickelectric_test-username",
+        unique_id="flick_electric_test-username",
     )
     entry.add_to_hass(hass)
 
     with patch(
-        "homeassistant.components.flickelectric.config_flow.SimpleFlickAuth.async_get_access_token",
+        "homeassistant.components.flick_electric.config_flow.SimpleFlickAuth.async_get_access_token",
         return_value="123456789abcdef",
     ):
         result = await _flow_submit(hass)
@@ -71,7 +71,7 @@ async def test_form_duplicate_login(hass):
 async def test_form_invalid_auth(hass):
     """Test we handle invalid auth."""
     with patch(
-        "homeassistant.components.flickelectric.config_flow.SimpleFlickAuth.async_get_access_token",
+        "homeassistant.components.flick_electric.config_flow.SimpleFlickAuth.async_get_access_token",
         side_effect=AuthException,
     ):
         result = await _flow_submit(hass)
@@ -83,7 +83,7 @@ async def test_form_invalid_auth(hass):
 async def test_form_cannot_connect(hass):
     """Test we handle cannot connect error."""
     with patch(
-        "homeassistant.components.flickelectric.config_flow.SimpleFlickAuth.async_get_access_token",
+        "homeassistant.components.flick_electric.config_flow.SimpleFlickAuth.async_get_access_token",
         side_effect=asyncio.TimeoutError,
     ):
         result = await _flow_submit(hass)
@@ -95,7 +95,7 @@ async def test_form_cannot_connect(hass):
 async def test_form_generic_exception(hass):
     """Test we handle cannot connect error."""
     with patch(
-        "homeassistant.components.flickelectric.config_flow.SimpleFlickAuth.async_get_access_token",
+        "homeassistant.components.flick_electric.config_flow.SimpleFlickAuth.async_get_access_token",
         side_effect=Exception,
     ):
         result = await _flow_submit(hass)
